Struct graphsearch::Node

pub struct Node<T> {
    pub content: T,
    pub adjacent: Vec<Vertex>,
}

A node in the graph, made up a any content type T and a Vec of vertices

Fields

content

content can be any type T

adjacent

adjacent takes a Vec of vertices to adjacent nodes