Struct graphsearch::Vertex

pub struct Vertex {
    pub cost: i32,
    pub node: usize,
}

A vertex between two Nodes with an associated i32 cost and a target node. Vertex derives Copy, Debug, Eq and PartialEq and implements Ord and PartialOrd as we use it ordered compound types.

Fields

cost

cost is defiened as an ì32, which might change

node

node, an usize index of the node at the other end of this vertex